+++ tests/e2e/config.py
@@ -1,4 +1,3 @@
-import datetime
 import os
 
 import pexpect
@@ -33,22 +32,3 @@ def shell_session(request):
     child.sendcontrol("D")
     child.close()
     child.logfile.close()
-
-
-def test_echo_command(shell_session):
-    shell_session.sendline('/bin/echo "Hello, World!"')
-    shell_session.expect("Hello, World!")
-    shell_session.expect(PROMPT)
-    assert "Hello, World!" in shell_session.before
-
-
-def test_ls_command(shell_session):
-    shell_session.sendline("/bin/ls -l")
-    shell_session.expect(PROMPT)
-    assert "total" in shell_session.before
-
-
-def test_pwd_command(shell_session):
-    shell_session.sendline("/bin/pwd")
-    shell_session.expect(PROMPT)
-    assert os.getcwd() in shell_session.before
